Improve containers and add tile integration
Containers are no longer items, and can now be used for general item
storage. ContainerItems are the new version of containers as items,
and will be used as such. Tiles have been re-added in a similar way to
Items and BaseItems. The only difference so far is that Tiles can
contain items.