Implement AsciiPanel functionality in Display
The Display class can now be constructed, allowing it to host an
AsciiPanel window. Added a Screen interface for display and input
processing. Added an ExtChars class to contain a list of extended ASCII
characters that can be used with AsciiPanel. Created an example screen
with display and input functionality. Main now launches a Display window
by default, though the former code is merely commented out. All new
classes are stored in the core.display package.