Overhaul class hierarchy and add more customizability to Terminal
Screen is now an abstract class with a Display field and more methods.
This means that each Screen will always have a Display associated with
it, and not just during output. Note that the Display can be changed,
though this still allows for more adaptable windows. Display is now the
class that should be passed around instead of AsciiPanel, due to its set
of utility functions.

Almost every part of the Terminal class can be customized now. This will
make it a much more flexible class in the future. A new Window class
will soon be made to sort out the more generic window functions from the
Terminal itself, which will then extend it.

Other Changes:
 -Display constructors no longer need a startScreen, and for security,
  an init() function has been added to keep risky functions out of
  constructors.
 -Point is now stored in core, and its x and y fields are now public,
  since the class is so simple.
 -Rename Display.terminal to Display.panel, since it the latter is more
  generic.
 -Rename TerminalScreen to Terminal now that there will be no confusion
  over it and Display.panel.
 -Fix the width parameter being ignored in Menu.printBoxed()
  functions.
 -Fix the inability to type the letter T in the terminal, since it was
  overridden by the switch in ExampleScreen.
 -Added Shift+Enter as a new cheat to square the score.