Separate Terminal and Window functionality
The Terminal had many useful, generic windowing functions built-in, so
they have been extracted to a new Window class. ExampleScreen (now
renamed to MasterScreen) now uses Windows as its subclassing, though
this will soon change as Window becomes an output class and Screens
return as a more general-purpose class.

Also changed the default panel height to 48. Most terminals are 80x24,
and since this system uses square characters, the 24 has been doubled so
that the physical size of the window is approximately the same.