Add the ability to create horizontal separators within Windows
Added lines with a drawing function in ShapeMaker and char storage in
LineBorder (name pending). null can be used as an array content, which
will be safely replaced by a separator in printBoxed() and
printCenterBoxed(). Empty lines can still be created by entering the
empty String "" into a content array.