Allow for more complex window separation
Windows can now be divided multiple times in any sequence of horizontal
and vertical separations. LineBorder now keeps track of whether the line
should be horizontal or vertical, and has new constructors to account
for this. Window contains a List of LineBorders, one for each separator,
and has utility functions for adding and changing these more easily.
Also fixed an issue with windows passing their x as topLine and y as
leftIndent; they now correctly pass x as leftIndent and y as topLine.

All of these changes can be seen in MasterScreen, which now uses a
Window instead of a PopupWindow for its output. Score is displayed on a
vertically-separated panel, and pressing Shift+Enter also adds a message
about the score being hacked, displayed in a new horizontal section.
Reimplemented movement of the window on the x-axis.