Improve NameGenerator, allowing it to handle many syllables
NameGenerator is no longer restricted merely to a prefix and suffix; it
now uses a two-dimensional array that can store many sets of syllables.
Many old methods no longer work due to this system, but better and more
generic methods have been added. A demo of this can be found, commented
out, in Main.