Add a heightmap generator using midpoint displacement
All of the midpoint displacement code was borrowed from Rudi van Etten
of Java Game Programming Step by Step. I reformatted it and adapted it
to the Map class, making it much more generic along the way. I also
added a way to exaggerate differences in elevation by taking the height
to a power and then expressing it as a fraction of the highest height.
I also added the heightmap to MapScreen, which can represent heights as
varying shades of gray. (Note that MapScreen is primarily for testing at
this point, and will eventually be at least cleaned up if not removed.)