Rework definition system to return a String[]
Items now return a String[] as their definition, which can be printed to
the Console or through a Display. An internal chain of defineAsList()
methods allow items to pass down their definitions as easy-to-modify
Lists of Strings. A demo of this can be found in Main. Also fixed the
Container class printing its used capacity over its total capacity,
leading to improper fractions.