Add Entity controls, scrolling, custom map display sizes
Rewrote the Entity movement system and implemented it in MapScreen with
a sample player. Added glyphs to Entities so they can be displayed, and
a List of Entities to Maps so that they can display them.

Maps may also be larger than the Display and still be displayed thanks
to the new exporting methods. toGlyphs() can show a range of Tiles and
Entities as their glyphs, which can be sized accurately to a Display.

Other Changes:
 -Added a toColorSetArray(ColorChar[][]) function to ColorSet
 -Added convertToIndex() and convertToCoordinates() methods in Map for
  quick conversion between the two location systems
 -Added a Map method to find a random tile with the given property;
  useful to prevent Entities from appearing inside walls