Update parameters of constructors in the item system
Item constructors used generic, single-character parameter names, which
could be quite confusing without source code or documentation. Now they
have all been updated to the this.field = field format, where field is
both the name of a field and the parameter assigned to the field. Also
made Console abstract, as it should be since it has no constructors.