Make AlignedWindow and PopupWindow extensions of Window
Window<Content> is now used to manage both existing types of window. The
old Window class is now called AlignedWindow since it is left-aligned.
Content in the new Window class refers to the type of each content; for
Windows this is ColorSet and for PopupWindow this is ColorString.

Also renamed Window.displayOutput() to Window.display() to avoid
confusion with Screen.displayOutput().