Split Terminal into PopupTerminal and AlignedTerminal
Similar to what was done with Window, Terminal has been customized for
both PopupWindows and AlignedWindows. AlignedTerminals are also capable
of displaying input with different colors. Better documentation for both
classes is pending, as well as more constructors for AlignedTerminal.

To achieve this update, add functions were added to ColorSet. These
allow it to function like a StringBuilder in some regards, such as
chaining additions.