Revise Tile system and update Map (now TileMap)
Remove the TileType system, since this will need to be customized as
constants for each project project. TileProperty is also gone, instead
being accepted as a Property parameter in Tile and TileMap. TileMap now
has an Iterator, which is yet unused in the library but may be used
elsewhere. It also stores its own RNG and has protected fields for any
classes extending it. Added toCosts() and toSquidGrid() methods, which
will convert the map into 2D arrays of doubles or chars, respectively.
toSquidGrid() uses SquidLib's definitions for walls and floors, with a
hash ('#') representing a wall and anything else representing a floor.

Other Changes:
 -Added equals() methods to Tile and ColorChar
 -Moved old Tile class to ItemTile
 -Improved Entity directional movement with the fields in Direction
 -Added an isOpen() method to Map, rooted in tile costs