Add a menu system, confirmation screen, and fix popup windows
Menu is a new class that acts as an extension of Window, though it is an
inheritance relationship. Each content of the Window can be selected
and cycled through using the directional keys. ColoredMenu is currently
the only implementation of Menu that shows which item is selected, which
it does by setting the foreground and background colors of the selected
item.

Added a confirmation system that abstracts common behavior of subscreens
to a class where they can be easily re-used. The methods onConfirm() and
onCancel() are designed to be overridden and customized for each object,
and are called when enter and cancel are pressed.

Also added several copying constructors to help create copies of objects
without editing objects passed in as references. Windows can now be
constructed with initial contents as well. Fixed the contents of popup
windows being hidden by the filled boxes surrounding them.