Move higher-level Entity functionality to Being and add AI class
Entities have been greatly simplified so that they can operate merely as
an object that is not tied down to a tile. Being has inherited most of
Entity's old functionality, including speed. A flexible AI class now
manages the behavior of Beings, and is designed to be overridden in all
implementations.