Replace most fail cases with exception throws
In many cases, methods returned null or exited early instead of properly
throwing an exception. Methods will now throw exceptions, some of which
are managed at higher levels of the program (for instance, Window's
subclasses will simply not display themselves if WindowBuilder throws an
exception). These changes have been the most profound for FileManager,
and have led to the removal of some outdated methods in Console and
Prompt. Also fixed the alignment of separators and borders when a column
of a Window is longer than a previous column.