Add audio methods and path manipulation methods to FileManager
FileManager now has methods used to play and loop audio files. These
accept WAV files, and have not been tested with any other file types.
More methods to manipulate the path have been added, including one to
move up to the containing folder and another to add another folder.
Since CoreLibrary is to be included in other projects as a jar and thus
stored in a lib folder, these methods will be necessary to get the path
to a valid directory.