Remove specific fields from Being
In an effort to make Being and AI more customizable, the speed and
inventories fields of Being have been removed. If these need to be
implemented in a project, then Being can be extended. In any case, AI is
probably a better candidate for managing speed and actions, and it was
designed to be extended initially.