Remove Display's static graphical methods
It is unlikely that an AsciiPanel would be used on its own outside of a Display, so the static graphical methods in Display were removed and their bodies given to the local methods. Also removed bookmark comments as they were unused and only IDE-based (meaning that it would be possible to create named bookmarks without comments).