Fix line and indentation issues in AlignedWindow
In each increment of curIndent (after a vertical separator), the previous value of curIndent had been added to curMaxLength and would be added back to curIndent. This has been fixed by resetting it with = instead of +=, and starting with the initial location. There were also cases in which horizontal separators wouldn't fill the entire window, and since they always should, this was easily overridden with a reference to overallMaxLength.