Fix Gradle and remove wrapper, fix deprecation
APWT can be built with an instance of Gradle installed on the system,
removing the need for the wrapper. This was tested with Gradle 7.3.3.

Fixes `build.gradle` to use `implementation` instead of `compile` for
dependencies, and uses fixed versions of dependencies that are known to
work. Also ensures compatibility with Java 8, because why not?

Also fixes a deprecation warning by changing `new Integer(int)` to
`Integer.valueOf(int)`.